requirements.txt
selenium==4.24.0
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6

